fix(cli): isolate terminal mouse parsing from React
