+++ apps/cli/src/ui/tui/mouse.ts
@@ -3,29 +3,10 @@
 import { useEffect, useRef } from "react";
 import { useStdin, useStdout } from "ink";
 
-export type MousePress = { readonly x: number; readonly y: number };
+import { mousePressesIn, type MousePress } from "./mouse-input.ts";
 
 const ENABLE_MOUSE = "\u001B[?1000h\u001B[?1006h";
 const DISABLE_MOUSE = "\u001B[?1006l\u001B[?1000l";
-const SGR_MOUSE = /\u001B\[<(\d+);(\d+);(\d+)([Mm])/gu;
-const INK_MOUSE_INPUT = /^\[<\d+;\d+;\d+[Mm]$/u;
-
-/** Ink passes an unknown SGR mouse sequence through without its Escape byte. */
-export function isMouseInput(input: string): boolean {
-  return INK_MOUSE_INPUT.test(input);
-}
-
-/** Complete left-button presses in one terminal input chunk. */
-export function mousePressesIn(input: string): readonly MousePress[] {
-  const presses: MousePress[] = [];
-  for (const match of input.matchAll(SGR_MOUSE)) {
-    const button = Number(match[1]);
-    if (match[4] !== "M" || (button & 3) !== 0 || (button & 96) !== 0) continue;
-    presses.push({ x: Number(match[2]) - 1, y: Number(match[3]) - 1 });
-  }
-  return presses;
-}
-
 /**
  * Report terminal clicks while mounted and restore the terminal mode on exit.
  * Terminals without SGR mouse support ignore the mode sequences; keyboard use
